refactor: deduplicate OAuthClientRow type, move OAuthClientForm to components
